backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.19
alembic>=1.11.1
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
google-genai>=0.2.0
PyPDF2>=3.0.0
python-docx>=0.8.11
